+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## [0.0.2]
+Complete rewrite.
+
+### Added
+- `config()` method
+- `matches()` method
+- `mayGrant()` method
+
 ## [0.0.1]
 ### Added
 - `verify()` method